Split channel mutations
