+++ saleor/graphql/channel/mutations/__init__.py
@@ -1,0 +1,17 @@
+from .base_channel_listing import BaseChannelListingMutation
+from .channel_activate import ChannelActivate
+from .channel_create import ChannelCreate
+from .channel_deactivate import ChannelDeactivate
+from .channel_delete import ChannelDelete
+from .channel_reorder_warehouses import ChannelReorderWarehouses
+from .channel_update import ChannelUpdate
+
+__all__ = [
+    "BaseChannelListingMutation",
+    "ChannelActivate",
+    "ChannelCreate",
+    "ChannelDeactivate",
+    "ChannelDelete",
+    "ChannelReorderWarehouses",
+    "ChannelUpdate",
+]
